refactor(core): implement type-safe array indexing for core types
Eliminates verbose `as usize` casts across the codebase by implementing
`Index`
and `IndexMut` directly for core enum types (`PieceType`, `Piece`,
`Side`, `Square`).

- Add `impl_index!` macro in `src/core/types/mod.rs` and apply it to
core types
- Remove manual `Index` implementations for `PieceType`, `Piece`,
`Side`, and `Square`
- Replace direct array index casts with type-safe enum indexing across
position logic,
  move generation, and search history
- Use `Square::mirrored()` in `piece_square_tables.rs` instead of
repeating
  manual rank/file vertical and horizontal mirroring arithmetic
- Simplify `compute_defender_bonus` to directly sum and return the
bonuses
- Mark query helpers in `queries.rs` as `const fn`

+++ src/search/history_moves.rs
@@ -19,21 +19,21 @@ impl HistoryMoves {
 
     #[inline]
     pub fn get(&self, color: Side, mv: Move) -> i32 {
-        self.table[color as usize][mv.from() as usize][mv.to() as usize]
+        self.table[color][mv.from()][mv.to()]
     }
 
     #[inline]
     pub fn increase(&mut self, color: Side, mv: Move, depth: i8) {
         let depth = depth as i32;
         let bonus = depth * depth;
-        self.table[color as usize][mv.from() as usize][mv.to() as usize] += bonus;
+        self.table[color][mv.from()][mv.to()] += bonus;
     }
 
     #[inline]
     pub fn decrease(&mut self, color: Side, mv: Move, depth: i8) {
         let depth = depth as i32;
         let bonus = depth * depth;
-        self.table[color as usize][mv.from() as usize][mv.to() as usize] -= bonus;
+        self.table[color][mv.from()][mv.to()] -= bonus;
     }
 
     pub fn decay(&mut self) {
